TM#setNetworkSelectionModeAutomatic hang muliple minutes
Root Cause:
TM#setNetworkSelectionModeAutomatic is implemented as a blocking
operation. When modem detects abnormal context (slow network
registration, temporary failure, lack of available network ect.),
it will wait for timout (which may be up to 200 seconds) before sending
telephony the result.

The will block the MainThreadRequest in PIM to return, although the API
is not designed to return any value.

Solution:
Introduces timeout for PIM#sendRequest and wait for maximum of 2 seonds
for TM#setNetworkSelectionModeAutomatic to return.

Bug: 176239119
Test: manual smoke test
Change-Id: Ic6be6c831d770845e0fa75dbfd23e72f61f7854b