Eliminate the interfere of removeMessage for dual-SIM
In dual-SIM case, when two SIMs are loaded and try to fetch
CarrierConfig, one call of removeMessage(code) will remove
the timeout messages of both subscriptions.

This is fixed by sending timeout msg with phoneId as msg.obj
and calling removeMessage(code, obj) to only remove timeout
msg with the matching phoneId.

Bug: 181979320
Test: Manual smoke test
Change-Id: Ie7b20a4f609ce1eb8039753ce976d16510879923