Add EXTRA_ANSWERING_DROPS_FG_CALL to incoming call in some DSDS cases.
When a waiting call comes in and there is an in-progress call on another
subscription, answering the incoming call will cause the ongoing call
to drop.

Detect this scenario and add EXTRA_ANSWERING_DROPS_FG_CALL to the waiting
call so that Dialer will notify the user of this scenario.

Bug: 178649879
Test: Added over complicated unit tests to verify extra adding behavior.
Change-Id: Id4caf8c210151bc0125b09a87e3c1dede2117649