Remove CallQualityManager reference in TelephonyConnection.
This code has been migrated elsewhere and it no longer required.
There was an issue that a BT call quality report could come in before the
underlying connection (hence phone and context) has been linked up to
the TelephonyConnection.  Since this code is no longer needed, removing
to fix the bug.

Test: Manual test to verify no NPE is present.
Fixes: 185857677
Change-Id: Ib3035c167238a9b6d58ea6d91c8cfd84950fde66