Correct issue with auto disconnect conference where no participants.
The conditional check was structured assuming that the number of
participants had to go from > 0 to 0.  However in the case where a single
party conference gets treated as a standlone call, there will be no
participants at the start.

The previous unit test case didn't catch this because it assumed that we'd
go from 2 participants to 0.

Test: Add new unit test case for this scenario.
Fixes: 184701944
Change-Id: I2a498d80f5f6ee2bef6b21d1520678d21fa9c1a7