Fix CDMA Call Forwarding/Call Waiting UI
The CDMA Call Forwarding and Call Waiting UI was incorrectly
being created in the CallFeaturesSettings. Since Call waiting requires
a TimeConsumingPreferenceListener, it can not be directly added to
the main CallFeaturesSettings activity. It should eventually follow
the GSM pattern (or be refactored to have one UI for both technologies).
This changes the call forwarding/call waiting settings to be under
"Additional Settings" in order to support the Call Waiting Activity.

Bug: 185963699
Test: manual; test UI
Change-Id: I981d9d825ba781b691b610ef741b1b2d01953317