ServiceStateProvider:Enforce location permissino check with targetSdkVersion R-
b/182384053 enforce location permission check for apps with targetSdkVersion
S+ to protect the sensitive info. However, for app with targetSdkVersion
R or lower, it does not enforce the same to keep maximum compatibility.

This patch will scrub the location sensitive info in
ServiceStateProvider if the app with targetSdkVersion R- without
location permission.

Bug: 191913126
Test: atest ServiceStateProviderTest ServiceStateTest
Change-Id: Ib30c05952bd1b3e3f7dbf239bbcf6388bf2ab9bd