Stash verstat in the connection extras.
This works around an issue where a RemoteConnection will not pass along
the verstat (there is not even an API for it).
To work around this we stash verstat in the extras and then in Telecom
we can check for this since remote connection services will pass all the
extras along unmodified.

Test: Manual testing on connection mgr carrier.
Bug: 192608283
Change-Id: I2377e6721179da5e022fc4011d77ef14c719fb71