Do not verify ImsService is connected for MMTEL user settings
Perform a partial revert of ag/14804216, which caused a RemoteException
to be thrown for the case where there is no ImsService up. This is a
regression in the API contract, so it must be partially reverted.

Instead, we will allow the user setting to be applied and the ImsService
will pick up the change when it comes back up.

Fixes: 192930411
Test: atest CtsTelephonyTestCases
Change-Id: I3aa06588af8391f61155663c6c6818d1bed4b89f