Make CDMA Call Options UT aware
Introduce the ability for CDMA call options to detect when SS
over UT becomes available or unavailable and integrate the SS
over CDMA fallback carrier config for situations where SS over
UT becomes unavailable.

1) In the case where SS over CDMA is configured, never disable the
supp service settings, because there will always be an avenue to
set the settings.
2) In the  case where SS over UT is available, but SS over CDMA is
not, take into account the availability of SS over UT when the
settings are loaded and enable/disable the preference based on that
state. If the device moves from SS over UT capable -> unavailable
during this time, the setting will be disabled. If it happens while
a pending request is already being serviced, the framework will
throw an exception causing the UX to handle this and show an error
as well as gray out the Preference.
3) in the case that neither SS over CDMA or UT is available, just
gray out the settings that require either one be available.

Bug: 191057045
Test: manual
Change-Id: Ie03cce86da911571f3b3b6c3563a2882861818d2