Fix IMS <-> CS redial scenarios for silent redial
Instead of always using the GsmCdmaPhone to register
for event callbacks on the TelephonyConnetion, let
telephony call back with a consumer indicating which
phone it is using. This allows us to set the
call event listeners properly.

Also, track the phone used to register events and
use that information to deregister events if a new
call to register events is performed

Bug: 193805172
Test: manual
Change-Id: I54a93b3acf9d56d343e4b254fcb11dd26a8f5d4a