Always get the latest Connection in the call for the incoming case
For the auto-missed call scenario, a disconnected connection may
still be tracked in the Call. If getEarliestConnection is used,
it will grab a potentially disconnected call vs the newest incoming
call.

Instead, always grab the latest connection, since this is most
likely going to be the connection that the incoming call notification
is for.

Bug: 193910479
Test: incoming call tests
Change-Id: Idc46344935357a7216768aea4579761d79bb51ec