Delay the continuous Call forward USSD commands
When the device receives the response of SET call forward command, then
the device starts the the GET call forward command. Since the frameworks
has the limitation, it can't handle two continuous USSDs if the AP sends
the commands too quickly. Therefore, the UI sets the delay about 1 sec.

Bug: 199244910
Test: build pass and request for the tester verify with this patch at
the bug.

Change-Id: I4bf21ebfa0674fef368dd3bc81d294e04bdd0ead