Delay the continuous call waiting USSD commands
When the device receives the response of SET call waiting command, then
the device starts the the GET call waiting command. Since the frameworks
has the limitation, it can't handle two continuous USSDs if the AP sends
the commands too quickly. Therefore, the UI sets the delay about 1 sec.

Bug: 199244910
Test: build pass and request for the tester verify with this patch at
the bug.

Change-Id: Ieacbda965aa99bf85be92b01035384fbc6a3787f