Add required flag to registerReceiver call in TelephonyConnectionService
Android T adds support to allow a runtime receiver to be registered as
not exported, but to ensure apps can take advantage of this, calls to
registerReceiver must specify a flag indicating whether the receiver
should be exported for apps targeting T+ that are registering for
non-system broadcasts. This commit adds the RECEIVER_EXPORTED
flag to the receiver in TelephonyConnectionService since this
broadcast is protected by a privileged|role permission and is
expected from the Dialer.

Bug: 161145287
Test: Build
Change-Id: I2a98fb8b5c2326c7fb35a90d63aa081287340195