Fix potential NPE in TelephonyConnection#refreshConferenceSupported
The old code could of caused a Null Pointer Exception when calling
getCarrierConfig to check if isRttMergeSupported.

It was discussed that the best action to take was to change
getCarrierConfig so that it is guaranteed a null will never
be returned.  The rest of the file was adjusted to cut down
on code bloat.

In summary functionality of TelephonyConnection remains the same.
More saftey measures have been put in place.

bug: 200104919
Test: 2 Unit Tests
Change-Id: Ie66db292c32c73ba26a4c2e7f780a6a36c6f11d1