Handle unsupported operations in the time consuming pref activity.
There are some requests which can be made (e.g. for Call barring) which
an operator will not support.  Rather than raising a "sim error" dialog
in this case, we will just disable the preference and NOT show an error
for a better user experience.

Also, moving the logs for CallBarringEditPreference out of the "DBG"
block and raising some to info.  It was hard to diagnose logs in this area
since most were missing or not available in standard info level logs.

Test: See other CL; this was tested manually by forcing the call barring
screen to be available on TMO.
Fixes: 231284401

Change-Id: I23d0929d04353077c8a22a12df0b97e6df3c9fd4