Move onRttInitiated callback onto TelephonyConnection handler.
In some cases it is possible for onRttInitiated to come in while a call
is being added to Telecom.  The call is added to Telecom on the main thread
handler, but the rtt initiated update comes in via a binder thread from
the IMS service.   As a consequence the RTT status is sent to Telecom and
THEN the new connection and its state is reported to Telecom.  This
overwrites the new rtt state previously reported.

Test: Manual regression testing for RTT initiation on live network.
Fixes: 234650328
Change-Id: I508a6960f7ace160ea787b2ea2c21a8d94d262e2