Suppress network selection not available notification
This change prevents the notification from showing when
the carrier config is overridden to hide the network selection
menu in Settings app.

"Network selection not available" notification is used to
guide user to re-select the available network when the
previously user selected network from Settings menu is
not available.

When the network selection is hidden by carrier apps,
the notification may confuse the end user. We should
suppress it in this case.

Bug: 241942743
Test: atest NotificationMgrTest (in ag/19632405)
Test: manual regression test (see go/network-selection-notif-test)
Change-Id: Ic042ba0b808c9045d41eb394001a1a82576b6fd8
Merged-In: Ic042ba0b808c9045d41eb394001a1a82576b6fd8