Added ignore modem/config flag to eccdata proto.
Updated the eccdata proto file to support the ability to set whether Modem/Config sourced emergency numbers should be ignored for each country. This flag will default to on (i.e. ignoring emergency numbers sourced from modem/config) for countries in which we have a high level of confidence in the accuracy of the Android Emergency Number Database, such as the US. If the value of this flag is set to true, we have a good authority of emergency numbers in the Android Emergency Number Database for this country.

Bug: 253101043
Test: Manual Test
Change-Id: I46da4ee9f504773de92c3adf49249ae84dbdb5d5