Talkback issue in Dialer Settings for back button
The voice prompt for the back button in dialer settings for Blocked
numbers, Calls, and Quick Responses displays "Navigate up" instead of
"Back". This is expected because it indicates that there is a view one
level higher where the aforementioned views were invoked from. Hence,
talkback assistant prompts user to "Navigate up (one level)".

To resolve the issue, we can provide an alternate content description
for back navigation by invoking setHomeActionContentDescription on the
given ActionBar instance.

Bug: 239758276
Test: Manual Test
Change-Id: I34cbbe4f17a4004cfce1fdb1045af7d28034e27f