Ensured all instances of ENT used for verification
Only one instance of
EmergencyNumberTracker was being used to verify emergency numbers. This
caused inaccuracies when multiple SIMs are active. This change refactors
that invocation  to ensure all instances of EmergencyNumberTracker are
used to verify emergency numbers via
TelephonyManager#isEmergencyNumber(). Added unit test to verify this
method is used when verifying emergency numbers.

Bug: 249116557
Test: atest TelephonyConnectionServiceTest#testIsEmergencyDeterminedByTelephonyManager
Change-Id: I4a635377cd29fac39c31a15b27a776a8dba29417