Add work profile support for registering phone accounts with Telecom
When registering a phone account for a user not associated with the
phone process, we want to grab the user associated with the
phone subscription. Currently, we enable MULTI_USER_CAPABILITY for all
phone accounts; we should disable this when registering another user.
We will also need the INTERACT_ACROSS_USERS permission to perform
actions across users (other than the primary).

Bug: 255342474
Test: PhoneUtilsTest.java
Change-Id: I78c2e879f0890cef09c3e95045bada4c5bbf1a01