Fix a race condition when registering IMS callbacks via ImsStateCallback
When the subscription is changed from an invalid subscription to a valid
subscription and ImsStateCallback#onAvailable() is called, the
application can register the MmTel capability or IMS provisioning
changed callback. However, the application can receive an exception
intermittently because the ImsManager instance for registering this
callback is different from the instance of ImsStateCallbackController.
To avoid this exception, the API registering the callback is changed to
use the ImsManager instance of the ImsStateCallbackController.

Bug: 265083810
Test: manual (verify registering capability/provisioning callbacks)
Change-Id: Ib803d70a8d06f4ad07d2c5e5afbdb165289b91a1