Support async void callbacks for SatelliteManager APIs
Instead of blocking the main thread until we receive a result from the
satellite service, call methods asynchronously and receive the result
through a callback. If a value was requested, return the value and error
through an OutcomeReceiver. If a value was not requested, return the
error through an integer consumer.

Test: atest SatelliteManagerTest
Bug: 268791334
Change-Id: I0cd345e14d242735d6be5b193e48bc22bdf69704