Update "useEmergencyPdn" flag for emergency SMS
This flag was introduced to check whether IMS PDN should be used for PS
emergency services. If IMS PDN is used for an emergency SMS, the domain
selection should not try to change the preferred transport type for
emergency service.
EmergencySmsDomainSelector sets this value based on the carrier
configuration according to whether the carrier supports the emergency
PDN or not when domain selection completes.

Bug: 270333880
Test: atest SmsDomainSelectorTest, EmergencySmsDomainSelectorTest
Test: manual (verify normal/emergency SMS on device)
Change-Id: Iea83306e26ec613e25ecaf3cccafb3d2b461fbcd