Ensure MERGE_COMPLETE connection event propagates for foreground call.
The connection event MERGE_COMPLETE was not propgating nor getting logged to the dumpsys as it was being invoked after the original calls had been diconnected in favor of the conference call and thus the listeners were removed at that point. This change ensures the MERGE_COMPLETE connection event properly propagates and logs to the dumpsys before the original foreground call is disconnected in favor of the conference call for IMS conferences.

Bug: 265187856
Test: Manually create an IMS conference call. Ensure MERGE_COMPLETE logs to the dumpsys for the original foreground call at time of merge.
Change-Id: I2426eddc88d49a9a6900ca65bb41d3295c86d772