Circumvent handler to ensure MERGE_COMPLETE propagates.
To ensure the MERGE_COMPLETE event logs before the listeners are removed, this CL circumvents the handler for MERGE_COMPLETE connection events and sends the connection event directly from TelephonyConnection#onConnectionEvent. This is intended to reduce flakiness in the ImsCallingTest suite and ensure MERGE_COMPLETE always propagates successfully for both the foreground and background call when an IMS conference call is merged.

Bug: 276799352
Test: atest ImsCallingTest#testConferenceCall
Change-Id: Ia944d11220e780432e7d3670258ec9a10f0202fe