Resolve user switch broadcast in account registry.
We saw a bug where forcing a manual crash on the phone process for the guest user results in the existing phone accounts being recreated with video capability enabled.

Do note that we already have built in logic to ensure that video
capilities aren't included when the current user isn't the admin. However, the broadcast is not received during boot up (since there are no user switches after the crash, no broadcasts would be received), and because mIsPrimaryUser is defaulted to true, Telecom incorrectly makes the assumption that the user is always going to be the admin.

The fix here would be to instantiate the value of mIsPrimaryUser to be whatever the current user is from ActivityManager, instead of defaulting to true.

Fixes: 286761407
Test: Manual to verify that mIsPrimaryUser is properly reflected when we (1) crash the phone process for both the admin/guest user and (2) crash the phone process for just the guest user.

Change-Id: I66269cf4669898b108e963a11ddf203eb2e501f2