DO NOT MERGE  Provide ability to place normal routed e-call on in-service sim first.
Handle the case where an emergency number has normal routing, and a
carrier wants to use the in-service SIM to place a normal routed
emergency call.  Ensures we'll prefer the in-service SIM in this case over
trying to use a limited service one.

Add '100' as a normal routed emergency number in India.

Test: testCreateOutgoingConnectionForNormalRoutedEmergencyCall - this
verifies the end-to-end case this bug was raised for and ensures that we
will prefer the in-service sim over the limited service sim.
Test: testIsNormalRoutedEmergencyCallForCarrier - this verifies the
new method isNormalRoutedEmergencyCallForCarrier to ensure it can properly
detect a normal routed emergency call for carriers that prefer to use an
in service sim.
Test: testIsAvailableForEmergencyCallsNotForCrossSim - regression tests
the method which was refactored and covers the new cases.
Test: testIsAvailableForEmergencyCallsForEmergencyRoutingInEmergencyOnly -
verifies that we can recognize whether a phone is available when the
phone is limited service.
Test: testIsAvailableForEmergencyCallsForEmergencyRoutingInService -
verifies that we can recognize a phone is available when the phone is in
service.
Test: testGetPhoneForNormalRoutedEmergencyCallWhenCarrierDoesntSupport -
verifies that if the carrier doesn't support using normal routing on the
in service phone we won't return anything.
Test: testGetPhoneForNormalRoutedEmergencyCallWhenCarrierDoesSupport -
verifies that if the carries does support using normal routing on the
in service phone we'll find it.
Test: testGetPhoneForNormalRoutedEmergencyCallWhenCarrierDoesSupportMultiSim
Similar to the latter but with multisim specifically.
Test: com.android.phone.ecc.EccDataTest#testEccDataContent
Bug: 281934614
Change-Id: I4151fb683b0df25a3bf0070ef71c90ba29ebf7cc