do not override the precise label for carrier defined busy tones
A partner reported a bug where the precise disconnect label was showing
as "Line Busy" instead of "Call Ended". This was a result of Telecom
overriding the precise disconnect label if the disconnect cause was in
the carrier config KEY_DISCONNECT_CAUSE_PLAY_BUSYTONE_INT_ARRAY.

This change ensures that the precise disconnect cause label will not be
overridden for busy tone values.

Bug: 296968778
Test: manual + 9 unit tests
Change-Id: Ie4232da73cdb66c056a8a4a585a807623c65cfaf