Fix a flaky test in ImsStateTrackerTest
Some of ImsStateTrackerTest are intermittently failed while processing
the posted messages. Since the posted messages are running first by the
thread switching, the expected result is different.
To avoid this situation, this change adds the message handling delay to
enable the thread switch between the posted message handling and current
execution flow.

Bug: 300853918
Test: atest -c TeleServiceTests:ImsStateTrackerTest --iterations 300
Change-Id: If64e1c50c6bf120b0de9794987a03f00df813332