Handle KEY_USE_ONLY_DIALED_SIM_ECC_LIST_BOOL
Usually, emergency number is identified based on all the active subscriptions,
no matter which subscription could be used.

If CarrierConfigManager.KEY_USE_ONLY_DIALED_SIM_ECC_LIST_BOOL is true
and the dialing number is not judged as an emergency number for the
given subscription on which dialing is placed, GsmCdmaPhone routes the dialing
as a normal call.

In that case, AP domain selection should do normal call domain
selection instead of emergency call domain selection.

Bug: 308863188
Test: atest TelephonyConnectionServiceTest
Change-Id: I991a103f1aa5d3bba7113c6e148a589fee13a862