Remove PL emergency numbers.
These extended emergency numbers are specified with "emergency" routing,
however in reality there is a per-carrier requirement for the numbers to
route via NORMAL routing or EMERGENCY routing.  Since the platform does
not support specifying the per-carrier routing, we will remove these
numbers for now.  The network and radio can still pass in the correct
numbers with routing appropriate to the active carrier.

Test: atest TeleServiceTests:EccDataTest#testEccDataContent
Bug: 301446552
Flag: This file is unflaggable and this is intended for 24Q1 release.
Change-Id: I6f7ef53dfd7833ea0b89de289ebc6b99d993e46d