Add SEND_EMERGENCY_INTENTS permission to Telephony's AndroidManifest.
This is needed because during an emergency call on a satellite-enabled device, Telephony sends a DISPLAY_EMERGENCY_MESSAGE ConnectionEvent to Dialer. In the Bundle of the ConnectionEvent is a PendingIntent. Dialer will launch this PendingIntent when the satellite button is pressed to launch the Satellite Questionnaire activity.

Without this permission, a "permission denied" error appears and the PendingIntent cannot be used. With this permission, the issue is resolved and Dialer is able to send the Intent.

Bug: 324485105
Test: built and deployed locally, tested that PendingIntent gets executed properly

Change-Id: I010893a8c1902f4131da9873c099b1cae2ed354c