Check the preferred scan type in limited service state
When selecting a domain in the initial state, if the UE is in the
limited service state, check the value of
KEY_EMERGENCY_NETWORK_SCAN_TYPE_INT.
If the value is SCAN_TYPE_FULL_SERVICE, do not select a domain and
request a network scan.
Otherwise, follow the existing logic.

Bug: 325185264
Test: atest EmergencyCallDomainSelectorTest
Change-Id: I4e493239eb119c3ecb379099a196d7b1389b85ea