Report anomaly when config change with subId maps to invalid phoneId
There is an increasing reports of exception from Pixels and other
OEM devices when CarrierConfigLoader received the notifyConfigChanged
with a subId.

The subId can pass the permission check but fail the following subId to
phoneId mapping check, which is pretty confusing.

To identify which scenario can cause such failure, an anomaly will be
reported to collect more context.

Bug: 270757342
Test: atest FrameworksTelephonyTestCases
Change-Id: If47cf7b23237177ea0368932e3389e4730eb82f0