Fixed incorrect SIM used for voice call
In some rare dual SIM scenarios, the voice call was
made on the wrong SIM. Fixed by always querying the
phone account for the default voice sub when we are
building the phone accounts. If phone account
information is out-of-sync between telephony and telecom,
then we set the correct phone account again.

Fix: 329049325
Test: Basic telephony functionality tests
Test: atest FrameworksTelephonyTests
Change-Id: I5c2b688237153c193cb6116c982d4cfd2ceb0a0e