Ensure CW button is disabled when it isn't enabled.
The call waiting UI incorrectly toggles (enables) the CW button when an
error occurs, indicating that it is enabled when that isn't case. When
the page is refreshed, the UI is updated accordingly but ideally, that
shouldn't have to occur on a refresh.

Bug: 328560292
Test: m Telephony (unable to test this manually since I don't have a SIM
that has CW enabled)

Change-Id: I14d04604774a1483870e89ed005d6a5489da0f1a
(cherry picked from commit 0af52d79d3bf884eab7ba136dda9265399a81076)