Select the phone in emergency callback mode for the new emergency call
While the device is in ECBM with the non-call preference SIM,
dialing another ECC is rejected by EmergencyStateTracker.

The phone for the call preference is selected for the new ECC.
However, since the phone for the call preference is different
from the phone with which in ECBM now, the call is rejected
by the EmergencyStateTracker.

Make TelephonyConnectionService select the phone in ECBM for
the new dialing.

Bug: 332543313
Test: atest TelephonyConnectionServiceTest
Manual test:
Precondition:
- Set the network to reject dialing request from SIM1.
- Initial Call preference: VZW SIM1 (pSIM)
- Initial DDS preference : VZW SIM2 (eSIM)
Test steps:
1. Turn on flight mode
2. Make an ECC (112/911)
3. End call
4. Make an ECC.
5. End call

Change-Id: I43a97b5fc62b9ae2488530a0ee22d10882b094b8