Keep PS preferred scan after VoLTE failure only once
ePDN setup can fail if it is requested while initial registration is
on going and the registration is rejected.
To recover this exceptional case, PS preferred scan is requested instead of CS preferred scan to retry on the same network with emergency attach.

However, this is required right after initial registration is
on going and the registration is rejected.
It means it's enough to try PS preferred scan after VoLTE failure only once.

So, make the PS preferred scan requested only once even with
the reason ImsReasonInfo.CODE_LOCAL_NOT_REGISTERED happens again.
Instead, do the RAT preference switching as usually.

Bug: 335752472
Test: atest EmergencyCallDomainSelectorTest
Change-Id: I87231443051c715b2be1aa06b668f1912c244389