PS preferred scan after VoLTE failure only when SIM ready
Keeping PS preferred scan after VoLTE failure is to recover
the exceptional case: ePDN setup fails since it is requested
while initial registration is on going and the registration
is rejected.

This case happens only when there is an active SIM.
So if there is no active SIM, do the RAT preference switching
as usually.

Bug: 335752472
Test: atest EmergencyCallDomainSelectorTest
Change-Id: Id72283ec4fd42022e6321868e8812ebe4678b9a2