Add DynamicRoutingController
Add a DynamicRoutingController that makes emergency calls through normal routing if normal calls are possible, and through emergency routing if not.
Add information to the resource about whether to use the dynamic routing feature, countries that use the dynamic routing feature, and emergency numbers that should be dialed through normal routing if possible.
Always use emergency routing if the source of the emergency number is Network or SIM.

Bug: 336759603
Test: atest TeleServiceTests
Test: build and manual test

Change-Id: I3660ff626085ea75365ac0bf51bcc39e943c1358