Check Wi-Fi network identification when monitoring connection state
Current implementation is based on the assumption onAvailabe of the 2nd
network shall be informed after receiving onLost of 1st network.
However, there can be cases that multiple Wi-Fi networks are available
at the same time. That is onAvailable of 2nd network can be notified
before receiving onLost of 1st network.

We need to distinguish the connection state per the network.
Remember the network identification and connection state respectively.

Flag: EXEMPT bugfix
Bug: 343611140
Test: atest EmergencyCallDomainSelectorTest
Change-Id: I494df98d785039e1001b89d6c5f5eb66de3ad308